Change Tasks to store usernames instead of Users
Storing Users was causing problems with object duplication in the
database, where each time a Task was saved with a User assigned, that
User would be readded to the database. To solve this, I replaced the
list of Users with a list of User references - the database uses
usernames as the User's unique identifier, so that is what gets stored
in the Task object.